mega/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
redis